# config/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "AlgorithemFinal",
      "version": "v1"
    },
    "paths": {
...
      "/api/Courses": {
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "",
          "parameters": [
            {
              "name": "search",
              "in": "query",
              "description": "قسمتی از نام",
              "schema": {
                "type": "string",
                "description": "قسمتی از نام",
                "nullable": true
              }
            },
            {
              "name": "unitCount",
              "in": "query",
              "description": "تعداد واحد",
              "schema": {
                "type": "string",
                "description": "تعداد واحد",
                "nullable": true
              }
            },
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CoursePaginatedResult"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Courses"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Courses/{id}": {
        "get": {
          "tags": [
            "Courses"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Courses"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Courses"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/Courses/{id}/TimeTables": {
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "برنامه زمانی درس هایی که این شناسه را دارند را برمیگرداند.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "شناسه",
              "required": true,
              "schema": {
                "type": "string",
                "description": "شناسه"
                
              }
            },
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTablePaginatedResult"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Courses/{id}/Masters": {
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "استاد های درس با شناسه وارد شده را بر می پرداند",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "شناسه",
              "required": true,
              "schema": {
                "type": "string",
                "description": "شناسه"
                
              }
            },
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MasterPaginatedResult"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Courses/{id}/Choose": {
        "post": {
          "tags": [
            "Courses"
          ],
          "summary": "استاد درس را برای ارایه انتخاب می کند. لازم که استاد را از طریق میدل ویر شناسایی کنید و از روی توکن",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "شناسه درس",
              "required": true,
              "schema": {
                "type": "string",
                "description": "شناسه درس"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "User": {
          "required": [
            "code"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "lastName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "code": {
              "type": "string",
              "description": "نام کاربری یا همان شماره دانشجویی"
            },
            "role": {
              "type": "string",
              "description": "master | admin | student",
              "nullable": true,
              "readOnly": true
            }
          },
          "additionalProperties": false
        },
        "Master": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "userId": {
              "type": "string"
              
            },
            "user": {
              "$ref": "#/components/schemas/User"
            }
          },
          "additionalProperties": false
        },
        "Day": {
          "required": [
            "label"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "label": {
              "type": "string"
            },
            "dayOfWeek": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "Bell": {
          "required": [
            "label"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "label": {
              "type": "string"
            },
            "bellOfDay": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "TimeTableBell": {
          "required": [
            "bell",
            "day"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "day": {
              "$ref": "#/components/schemas/Day"
            },
            "bell": {
              "$ref": "#/components/schemas/Bell"
            }
          },
          "additionalProperties": false
        },
        "Course": {
          "required": [
            "title"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "title": {
              "type": "string"
            },
            "unitsCount": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "TimeTable": {
          "required": [
            "course"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "master": {
              "$ref": "#/components/schemas/Master"
            },
            "timeTableBells": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTableBell"
              },
              "nullable": true
            },
            "course": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "additionalProperties": false
        },
...
        "CoursePaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Course"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "CourseRequest": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "nullable": true
            },
            "unitsCount": {
              "type": "string"
              ,
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "TimeTablePaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTable"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "MasterPaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Master"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
...
      },
      "securitySchemes": {
        "Token": {
          "type": "apiKey",
          "description": "JWT Authorization header using the Bearer scheme.",
          "name": "Authorization",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "Token": [ ]
      }
    ]
  }